fixed react doctor picks
