+++ frontend/components/markdown-renderer.tsx
@@ -58,19 +58,6 @@ const cleanupTableEmptyCells = (text: string): string => {
     .join("\n");
 };
 
-/**
- * Derives a display filename from citation data.
- * Extracts the filename from a file path or uses the filename field directly.
- */
-export const deriveDisplayFilename = (
-  filePath: string | undefined,
-  filename: string | undefined,
-  fallback: string = "Document",
-): string => {
-  const path = filePath || filename || fallback;
-  return path.split("/").pop() || path;
-};
-
 export const MarkdownRenderer = ({
   chatMessage,
   className,
